Move the creation of credential to the module
