+++ core/src/main/java/google/registry/tools/RegistryCli.java
@@ -31,7 +31,6 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterables;
 import google.registry.config.RegistryConfig;
 import google.registry.model.ofy.ObjectifyService;
-import google.registry.persistence.transaction.CloudSqlCredentialFactory;
 import google.registry.persistence.transaction.TransactionManagerFactory;
 import google.registry.tools.AuthModule.LoginRequiredException;
 import google.registry.tools.params.ParameterFactory;
@@ -239,8 +238,6 @@ final class RegistryCli implements AutoCloseable, CommandRunner {
       // Enable Cloud SQL for command that needs remote API as they will very likely use
       // Cloud SQL after the database migration. Note that the DB password is stored in Datastore
       // and it is already initialized above.
-      CloudSqlCredentialFactory.setupCredentialFactory(
-          component.googleCredentialJson(), component.googleCredentialScopes());
       TransactionManagerFactory.setJpaTm(component.nomulusToolJpaTransactionManager());
     }
